Autenticación y autorización listo. Falta agregar JS

+++ views/materias/index.php
@@ -1,11 +1,18 @@
 <?php require __DIR__ . "/../layouts/header.php"; ?>
 <div class="container mt-5">
-    <h1 class="text-center mb-4">Listado de Materias</h1>
+    <div class="d-flex justify-content-between align-items-center mb-4">
+        <h1 class="text-center mb-4">Listado de Materias</h1>
+        <div>
+            <span class="me-3">Bienvenido <strong><?= htmlspecialchars($_SESSION['usuario_nombre'])?></strong></span>
+            <a href="index.php?action=logout" class="btn btn-outline-danger btn-sm">Cerrar sesión</a>
+        </div>
+    </div>
 
 
+    <?php if(isset($_SESSION['usuario_rol']) && $_SESSION['usuario_rol'] === 'admin'): ?>
     <a href="index.php?action=crear" class="btn btn-primary mb-3">Agregar Nueva Materia</a>
     <br><br>
-
+    <?php endif; ?>
     <table class="table table-striped table-bordered table-hover">
         <thead class="table-dark">
             <tr>
@@ -13,7 +20,9 @@
                 <th>Materia</th>
                 <th>Año</th>
                 <th>Estado</th>
+                <?php if(isset($_SESSION['usuario_rol']) && $_SESSION['usuario_rol'] === 'admin'): ?>
                 <th>Acciones</th>
+                <?php endif; ?>
             </tr>
         </thead>
         <tbody>
@@ -26,12 +35,15 @@
                 echo "<td>" . $materia['mat_anio'] . "</td>";
                 echo "<td>" . $materia['est_nombre'] . "</td>";
                 // botones para eliminar y editar.
-                echo "<td>";
-                echo "<a href='index.php?action=editar&id={$materia['idMateria']}' class='btn btn-warning btn-sm me-2'>Editar</a>    ";
-                echo "<a href='index.php?action=eliminar&id={$materia['idMateria']}' class='btn btn-danger btn-sm'>Eliminar</a>";
-                echo "</td>";
+                if(isset($_SESSION['usuario_rol']) && $_SESSION['usuario_rol'] === 'admin'){
+                    echo "<td>";
+                    echo "<a href='index.php?action=editar&id={$materia['idMateria']}' class='btn btn-warning btn-sm me-2'>Editar</a>    ";
+                    echo "<a href='index.php?action=eliminar&id={$materia['idMateria']}' class='btn btn-danger btn-sm'>Eliminar</a>";
+                    echo "</td>";
+    
+                }
                 echo "</tr>";
-            }
+           }
             ?>
         </tbody>
     </table>
